requirements.txt
flask
pymysql
passlib
python-dotenv